details for singleplayer records also very small refactoring (I will redo Tetra Stats from the ground one day...)
