+++ lib/utils/relative_timestamps.dart
@@ -1,4 +1,9 @@
+import 'package:intl/intl.dart';
+import 'package:tetra_stats/gen/strings.g.dart';
 import 'package:tetra_stats/utils/numers_formats.dart';
+
+final NumberFormat secs = NumberFormat("00.###", LocaleSettings.currentLocale.languageCode);
+final NumberFormat _timeInSec = NumberFormat("#,###.###s.", LocaleSettings.currentLocale.languageCode);
 
 /// Returns string, that represents time difference between [dateTime] and now
 String relativeDateTime(DateTime dateTime){
@@ -56,4 +61,16 @@ String relativeDateTime(DateTime dateTime){
   } else {
     return inPast ? "${f1.format(timeInterval)} seconds ago" : "in ${f1.format(timeInterval)} seconds";
   }
+}
+
+/// Takes number of [microseconds] and returns readable 40 lines time
+String get40lTime(int microseconds){
+  return microseconds > 60000000 ? "${(microseconds/1000000/60).floor()}:${(secs.format(microseconds /1000000 % 60))}" : _timeInSec.format(microseconds / 1000000);
+}
+
+/// Readable [a] - [b], without sign
+String readableTimeDifference(Duration a, Duration b){
+  Duration result = a - b;
+  
+  return NumberFormat("0.000s;0.000s", LocaleSettings.currentLocale.languageCode).format(result.inMilliseconds/1000);
 }
